vmctl: merge influx2 into influx package, fix Next() record loss and query escaping
